pyproject.toml
[project]
name = "cookiebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ezcord>=0.7.4",
    "py-cord>=2.7.0",
    "python-dotenv>=1.2.1",
]
